Facelift 2024
- Added notifications feature
- Modified some styling
