+++ generator/templates/core/notification.ejs
@@ -1,0 +1,5 @@
+<div class="notification">
+  <p class="notification-title">📣 <%- title %></p>
+  <p class="notification-description"><%- description %></p>
+  <p class="notification-date">Published on: <%- date %></p>
+</div>
