+++ generator/templates/main.ejs
@@ -7,6 +7,11 @@
       <div class="content">
         <img src="https://sdk.bitmoji.com/render/panel/a7161b2e-c41a-4b71-a610-4dbb0184c2be-dae924a3-30cf-4be0-898e-39939da96209-v1.png?transparent=1&palette=1" class="u-block-center"></img>
         <p>Welcome to Jakes Documentation - This static site contains all the documentation that Jake has saved on their machine for various different pieces of work. Certain documents can be viewed within the site or if not supported be downloaded externally.</p>
+        <% if (notifications.length > 0) { %>
+          <% notifications.forEach(notification => { %>
+            <%- include('core/notification', notification) %>
+          <% }); %>
+        <% } %>
       </div>
     </div>
   </body>
